refactor(sqlx): continue migration - unified_user_handlers progress
- Converted QueryableByName derives to sqlx::FromRow
- Removed #[diesel(sql_type = ...)] attributes
- Migrated diesel::sql_query to sqlx::query/query_as
- Replaced get_result with fetch_one/fetch_optional
- Replaced load with fetch_all

+++ apps/backend/src/infrastructure/cqrs/projections/wallet_projection.rs
@@ -7,7 +7,7 @@ use crate::infrastructure::cqrs::projection::{Projection, ProjectionCheckpoint, 
 use crate::prelude::*;
 use async_trait::async_trait;
 use chrono::Utc;
-use sqlx::{PgPool, Postgres};
+use sqlx::{PgPool, Postgres, Transaction};
 use std::sync::Arc;
 use uuid::Uuid;
 
@@ -74,8 +74,8 @@ impl Projection for WalletReadModelProjection {
             ProjectionCheckpoint {
                 projection_name: self.projection_name().to_string(),
                 last_processed_event_id: event_id_text.and_then(|s| Uuid::parse_str(&s).ok()),
-                last_processed_sequence: seq as u64,
-                events_processed_count: count as u64,
+                last_processed_sequence: seq,
+                events_processed_count: count,
                 processed_at: Utc::now(),
                 is_healthy: healthy,
             }
